+++ usage/app/profile/profile-form.tsx
@@ -1,4 +1,10 @@
-import React, { memo, useCallback, useEffect, useMemo } from 'react';
+import React, {
+  BaseSyntheticEvent,
+  memo,
+  useCallback,
+  useEffect,
+  useMemo,
+} from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Form, validateAll } from '@/src';
 import { EFormName } from '../app-enums';
@@ -10,7 +16,7 @@ import { profileFormValidator } from './profile-form-validator';
 export const ProfileForm = memo(() => {
   const dispatch = useDispatch();
 
-  const fullName = useSelector((state: TAppState): string => state.response.fullName);
+  const fullName = useSelector((state: TAppState): string => state.response.fullName ?? '');
 
   const initialValues = useMemo(() => ({
     [EProfileFieldName.FullName]: fullName,
@@ -21,10 +27,12 @@ export const ProfileForm = memo(() => {
       setTimeout(() => {
         resolve('Oleg Mukhov');
       }, 300);
-    }).then((fullName: string) => dispatch({ type: 'SET_RESPONSE', payload: fullName }));
+    }).then((fullName: string): void => {
+      dispatch({ type: 'SET_RESPONSE', payload: fullName });
+    });
   }, []);
 
-  const onSubmit = useCallback((event) => {
+  const onSubmit = useCallback((event: BaseSyntheticEvent) => {
     event.preventDefault();
     dispatch(validateAll(EFormName.Profile));
   }, []);